Request YouTube to give us English video titles
Explicitly requests YouTube to give us a video title in "en" locale
if such a translation is available on their end, and uses the localized
title instead of the default title. Does not do anything for the case
that the first YouTube API call fails and the secondary node module
for grabbing youtube video information is called.